Fix compiler warnings and errors in ldb.c
Summary:
stdlib.h is needed for exit()
--readhead --> --readahead

Task ID: #

Blame Rev:

Test Plan:
compile

Revert Plan:

Database Impact:

Memcache Impact:

Other Notes:

EImportant:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -
fix compiler warnings & errors

Reviewers: dhruba

Reviewed By: dhruba

CC: heyongqiang

Differential Revision: https://reviews.facebook.net/D5805

+++ tools/ldb.cc
@@ -5,6 +5,7 @@
 #include <string>
 #include <iostream>
 #include <sstream>
+#include <stdlib.h>
 
 #include "leveldb/db.h"
 #include "leveldb/options.h"
@@ -122,7 +123,7 @@ static void dump_db(
     }
   }
 
-  fprintf(stdout, "Keys in range: %d\n", count);
+  fprintf(stdout, "Keys in range: %lld\n", (long long) count);
 
   // Clean up
   delete iter;
